+++ CHANGES.md
@@ -8,6 +8,11 @@ Version 1.3.6
 
 To be released.
 
+ -  Fixed a bug where `getUserAgent()` function had returned a `User-Agent`
+    string with a wrong JavaScript runtime name on Node.js.  [[#203]]
+
+[#203]: https://github.com/fedify-dev/fedify/issues/203
+
 
 Version 1.3.5
 -------------