+++ CHANGES.md
@@ -6,7 +6,7 @@ Fedify changelog
 Version 1.3.6
 -------------
 
-To be released.
+Released on January 31, 2025.
 
  -  Fixed a bug where `getUserAgent()` function had returned a `User-Agent`
     string with a wrong JavaScript runtime name on Node.js.  [[#203]]